added join to create functionality

+++ events/client/voiceStateUpdate.js
@@ -1,32 +1,40 @@
+const joinToCreateSchema = require("../../models/joinToCreateSchema");
+
 module.exports = {
     name: "voiceStateUpdate",
 
     async execute(client, oldState, newState) {
+        const schema = await joinToCreateSchema.findOne({ guildId: oldState.guild.id || newState.guild.id });
+
+        if (!schema.enabled) return;
+
         if (client.temporaryVoiceChannels.get(oldState.channelId) && !oldState.channel.members.size) {
             oldState.channel.delete();
-
             return;
         }
 
-        if (newState.channelId !== "929170092435054643") return;
+        if (newState?.channelId === schema?.channelId) {
+            const { guild, user, voice, id } = newState.member;
 
-        const { guild, user, voice, id } = newState.member;
+            const parent = newState.channel?.parentId;
+            const parentId = parent ? { parent } : {};
 
-        const vc = await guild.channels.create(`${user.username}'s channel`,
-            {
-                type: "GUILD_VOICE",
-                parent: "913854802121883698",
-                permissionOverwrites: [
-                    {
-                        id: id,
-                        allow: ['MANAGE_CHANNELS'],
-                    }
-                ]
-            }
-        );
+            const vc = await guild.channels.create(`${user.username}'s channel`,
+                {
+                    type: "GUILD_VOICE",
+                    ...parentId,
+                    permissionOverwrites: [
+                        {
+                            id: id,
+                            allow: ['MANAGE_CHANNELS'],
+                        }
+                    ]
+                }
+            );
 
-        client.temporaryVoiceChannels.set(vc.id, newState.member);
-        voice.setChannel(vc.id);
+            client.temporaryVoiceChannels.set(vc.id, newState.member);
+            voice.setChannel(vc.id);
+        };
     }
 };
